maybe fix bug with two USBs on SelectProfile
This is difficult to debug because my Linux VM seems to keep changing the by-path paths each reboot I do.

At the very least, it should be possible to proceed past this screen when two USBs profiles are in use.  It is unclear whether both are actually being loaded and used properly, however.