minor cleanup to SSMCasual
I noticed during mute's twitch livestream that the right edge of the glowing border around the song jackets wasn't showing (though, it was on my laptop).  This commit fixes a rounding error that caused that.

This also adds Notice Me Benpai and TLOES Chapter 1 to CasualMode-Groups.txt because both packs have full difficulties and are likely to be on many machines.