only force choices in SSMCasual if timer runs out
There was previously an oversight with the MenuTimer in SSMCasual.

Default steps and mods were always being forced because the the engine sets the MenuTimer to 0 just before transition the screen out, and I was checking to see if the MenuTimer was 0 to decide whether to force steps/scrollspeed or not!

It should be fixed now.