Remove ECFA hardcoding from FAplus flag
Now, if users modify the metrics so that ECFA white fantastic scores less than ECFA blue fantastic,
SubtractiveScoring will now behave properly, i.e. it still total up missed DP on white fantastics.
I did this a while ago for StomperZ, but for some reason never changed it for ECFA.

Also fixes an oversight with SubtractiveScoring so that it's centered when possible. (to keep symmetry.)