"Cover" -> "Hide BG" in en.ini for ScreenEval
Someone asked what "Cover" did after seeing it appear in a screenshot of ScreenEvaluation.  This was a pretty reasonable question since in PlayerOptions, the modifier was referred to as "Hide Background"!

This commit adds (another) special case to the routine that builds a string list of PlayerOptions for ScreenEvaluation to convert "Cover" into a more meaningful string value ("Hide BG" in en.ini).

This suffices for now, but if more special cases arise in the future, I'm going to have to rethink how I'm doing this.