increase number of MusicWheel items
Some of the recent UI changes (more transparency of certain elements in RainbowMode) revealed that MusicWheelItems could be seen suddenly appearing at the bottom of the screen, under PLAYER_2's PaneDisplay.

This commit resolves by increasing the number of MusicWheelItems drawn from 14 to 15.  Hopefully this doesn't have unintended consequences...

This also adds code to cut the SSM MusicWheel "cascade" animation out of the draw routine once it is done.