correctly get BPM ranges for autogenerated Courses
The previously implementation recognized that courseEntry:GetSong() would always return nil for autogenerated Courses, but didn't do anything to handle it.  This resulted in ScreenPlayerOptions throwing Lua errors when attempting to find the BPM range for autogenerated Courses.

With this commit I found that, for whatever reason (oversight in the engine?), trailEntry:GetSong() correctly returns the song object, even if an autogenerated course was chosen.  With this, I was able to still retrieve the BPM range of a particular Trail within an autogenerated Course, which is good enough!

Keep in mind that a "Course" is what both players choose, but that each player can then choose a distinct "Trail" within that Course (as long as LockCourseDifficulties=0).