remove problematic attacks from ScreenAttackMenu
It seems that attempting to enforce a turn mod (left, right, mirror, etc.) causes StepMania to crash, but only in Linux.

  • Having turn mods in Attacks causes SM to crash in Linux.
  • Removing those turn mods results in SM5 not crashing in Linux.
  • macOS and Windows seem to just ignore turn mods as attacks.

In my Linux VM and on the PSU dedicab, the arrows go past the receptors without judgment occurring for a little while, then SM5 crashes.  The crashlogs suggest it is audio related (RageSoundDriver_Generic_Software.cpp:204 ?).

I'm not entirely sure what's going on, to be honest.

My hunch is that turn mods aren't supported as Attacks because the underlying data structure that contains what columns are used in what sequence in Gameplay can't change on the fly.  But that's just a guess; I should try to verify it via src, if possible.

In any event, I'm removing turn mods from ScreenAttackMenu for the time being to prevent players from unwittingly using them.

I am also opting to remove the SpeedMod and SpeedModType OptionRows from ScreenAttackMenu as their presence caused the screen to always apply some speedmod (1x by default) as an attack, when players probably rarely wanted this...