Fix judgment graphics handling
Remove stray spaces from Competitive graphics filenames.

Work around an engine bug using a display kludge to turn underscores into spaces, and backticks into periods, in the display names for the judgment font option row. This allows judgment font names 'A', 'A_B' to coexist in the same folder, but display as 'A', 'A B'. If stored on the filesystem in the latter form, and we try to load 'A', SM will throw a bizarrely harmless but rather unhelpful error complaining that 'A B' is a duplicate of 'A'. I suspect this has to do with how spaces are parsed in loading bitmaps and bitmap hints.

That also fixes the 3.9 judgement font, which previously worked only serendipitously.

Some more obligatory code quality cleanup changes in/around the touched zones.