scroll NPS density graph for very long songs
I realized at ECS 6 last year that the density histogram became increasingly less helpful the longer the song is.  For songs longer than, say, 8 minutes in duration, it had to fit too much visual information (the x-axis of the graph) in too small a horizontal space.  It was basically useless for hour-long songs.

Thus, I'm changing the behavior of the graph to only visualize 4 minutes of the song at a time and scroll with the song to reveal the next 3 minutes of what lies ahead.

---

One of the inline comments bears repeating here:

The "graph_mask" here is not a "mask" in the proper sense.  I did experiment with that briefly, but abandoned it because it impacted the rendering of many ITG NoteSkins that also use masks...

The graph_mask is really just a black quad positioned in front of the density_graph_and_lifeline but behind all other underlay elements (like danger and subtractive scoring and etc.).  Once the DesityGraph starts scrolling, this quad will block it, effectively serving as a mask.

---

I've tested this with:

	• both players
	• using short songs and long songs and VERY long songs
	• in CourseMode and in regular gameplay
	• using 16:9 and 4:3 displays

and I'm SURE there are still bugs somewhere in this.

If you're reading these notes and are able to test this feature on your setup, please do and let me know if you find bugs.