cycle through Credit, Description, ChartName
People are apparently now using .ssc metadata correctly/appropriate and filling out chart Credit, Description, and ChartName fields uniquely.  I don't have any such .ssc files to verify this, but I'll take the community's word on it.

Thus, this commit changes the behavior of the StepArtist BitmapText actor on ScreenSelectMusic and ScreenEvaluation.  It would previously only display steps:GetCredit() (or course:GetScripter() in CourseMode).  The behavior has now been changed to look up

	• steps:GetCredit()
	• steps:GetDescription()
	• steps:GetChartName()

in "normal" mode, or

	• course:GetScripter()
	• course:GetDescription()

in CourseMode, and "marquee" through them, displaying one at a time for 2-second intervals.  Empty strings are ignored and not cycled through.

I've done a small amount of testing for this, but someone with newer/actual content should verify that this does what the community wants.