handle AutoStyle bug (sort of)
This commit does two things:

1. It causes screen branching to ignore AutoStyle when CoinMode is Pay. If CoinMode is Pay, players should always be directed to ScreenSelectStyle to manually choose a style.

This makes sense in the context of a public arcade (that might have accidentally set the AutoStyle setting) while still allowing the setting to exist for private machine owners and keyboard players.

2. If AutoStyle is set to 1 Player, and CoinMode isn't Pay, and both players have already been joined on ScreenTitleJoin, forcibly unjoin PLAYER_2 to prevent possible theme errors like "too many players joined for single" or "no style set" and etc.

Hardcoding the theme to unjoin PLAYER_2 isn't a great fix, but I'm not sure how else to handle the problem at this time.