fix oversight with preventing sorting into Double
e388059312a4306cb34397d5e484267fcedc6952 added code to check the Premium preference to see whether sorting into Double should be allowed or not.

Unfortunately, this was slightly too blunt of a fix, as an at-home player might never have set their Premium setting, and then be confused why they were unable to sort in double while on freeplay or home mode.

This adds an extra condition to only remove sorting into double if we're both in pay mode and Premium is off.