add TotalNotesHitThisGame stat to ScreenGameOver
Tally the total number of non-Miss judgments that occurred during a gameplay session and display them on ScreenGameOver.

This was requested.  It's probably more useful and interesting than displaying the player's current combo at ScreenGameOver, so this replaces that.

This will be inaccurate if any of the songs played during the session had custom, non-1 #COMBOS parameters set.  Only English is supported for now.

soft pliable clay
we lose form upon embrace
my shouting wakes me