fix oversights with TotalNotesHitThisGame
This fixes two oversights with the previous commits that added code to ScreenGameOver to display the TotalNotesHitThisGame per player.

The previous code would just get the total number of tapnotes in a stepchart and subtract the number of Miss judgments found in the per-colum data.  This was flawed in that it would still use that same total number of tapnotes even if a player quit out of a song early.  The new code just sums all non-Miss per-column judgments.

Also, the previous code did not handle the possibility of either player late-joining and would throw Lua errors if that happened.  That's fixed now, too.

the message contains
apologies just in case
I don't reach you first