fix DecentsWayOff bug on multipage Eval Summary
It's possible for players to change the DecentsWayOffs settings from song to song.  There was previously a bug in the ScreenEvaluationSummary code where the numbers for Decents and numbers for WayOffs would stay hidden once visible(false) was set.  This was problematic for multipage EvaluationSummaries.

This change ensures that the BitmapText actors for the numbers are appropriately reset to visible(true) every time the player switches from one page of scores to another.