don't include unavailable songs in group length
Commit 88190460a76060c27b5de76eb69065d23d85f146 added some code to calculate overall group length and display it when the player was hovering on the pack title in the MusicWheel.  It bluntly went through all songs in a given group and summed their lengths, regardless of whether the song currently appeared in the MusicWheel or not due to Long and Marathon cutoffs.

With this commit, we now consider the song's "cost" with regard to the number of stages remaining to only sum available songs.  This has been lightly tested and works with the Continue system.

What about hidden songs?

If players have HiddenSongs=1 set in their Preferences.ini, songs with #SELECTABLE:NO; in their .sm or .ssc file won't appear in SL's MusicWheel but will still be counted toward the overall pack length.  SM5's Lua API doesn't currently have a method for accessing a song's SELECTABLE tag, so the only recourse would be to parse the simfile – possible but not worth the effort right now.