always display BPM on ScreenEval
As requested in https://github.com/ddrillini/Simply-Love-SM5/issues/24

There was previously code to display the BPM range and music rate on ScreenEval, but it would only appear if MusicRate ~= 1.0, in accordance with hurtpiggypig's original design.

This has been requested repeatedly and I'm mostly indifferent.  I think it is unnecessary but its constant presence doesn't kill the existing aesthetic.  I did visually shift and resize elements very mildly.

---

Pittsburgh was darkness.  The taste of smoke in one's throat and heavy smog and black soot.  That was Pittsburgh.