hopefully fix bug with similarly named judgments
Commit 2cddbb0c0aef63ed441cf959472d4ac3bd4a36d7 by @andrewipark made a solid effort to fix the bug where the StepMania engine would complain about similarly named judgment graphics (for example, "Love" and "Love Chroma") and not load either.

The commit definitely cleaned up the relevant code into nice functions and clarified many messy comments originally authored by me, but it

1. required players to change judgment filenames – players were instructed to use underscores in place of whitespace ("DDR SuperNOVA 2x6.png" became "DDR_SuperNOVA 2x6.png"), and backticks in place of dots ("3.9 1x6.png" would become "3`9 1x6.png")

2. didn't always seem to work.  The engine still complained about multiple files matching "Love" if players had "Love" and "Love_Chroma" both installed.  See: https://github.com/ddrillini/Simply-Love-SM5/issues/39

I think I've fixed it.  GetJudgmentGraphics() would previously return a table of judgment graphic names like

{ "Love", "Love Chroma", "3.9" }

and it now returns a nested table like

{
   { "Love", "Love 2x6.png" },
   { "Love Chroma", "Love Chroma 2x6.png" },
   { "3.9", "3.9 1x6.png" },
}

By supplying both the user-facing string as well as the filename, we can ensure that LoadActor() doesn't get confused when loading the judgment sprites.  This allowed me to remove the underscore and backtick kludge as well.

I encountered only one oddity during my testing – the "3.9" judgment graphic would never load from Profile correctly.  It turned out that IniFile.ReadFile() was automatically converting it to a number which broke later string comparison.

I've special-cased a fix into SL-CustomProfiles.lua for now.

This has been tested on macOS but should be tested on Windows prior to any sort of release.

---

On the night of 2 December, arsonists made an unsuccessful attempt to burn the Natural Science Club's goat.  On 17 December, a Swedish news site reported that one of the guards tasked with protecting the Southern Merchants' goat had been offered payment to leave his post so that the goat could be stolen via helicopter and transported to Stockholm.  Both goats survived and were dismantled and returned to storage in early January 2011.