(attempt to) highlight HighScore rows in EventMode
In not-EventMode, GetMachineHighScoreIndex() will return the index of the player's highscore within the HighScoreList for the current stepchart for the current song.  I need to loop through that HighScoreList to generate a bunch of rows of BitmapText actors on Eval's Pane3 and NameEntryTraditional, so I normally just compare the loop index against the HighScoreIndex to determine whether to light up the current row.

Calling GetMachineHighScoreIndex() on a PlayerStageStats object in EventMode, however, always returns -1.  (Why?)

A wildly roundabout check is can be used in the case of EventMode to attempt to light up the correct highscore row for Eval's Pane3 and NameEntryTraditional.

The old logic recognized that EventMode always got -1 as indices but returned a ton of false positives (when players tied their scores, or when a new score tied an existing score already in the highscorelist) and frequently resulted in multiple rows lighting up.

The new logic won't return any false positives, but will return false negatives in extreme circumstances resulting in no HighScore rows lighting up.  Oh well.

"Extreme circumstances" are when we're in EventMode and both players earn a HighScore and they are both tied in score and neither is using a profile.

Overall, this should be better.

---

"That's the secret of how to enjoy writing and how to make yourself meet high standards," said Mrs. Berman. "You don't write for the whole world, and you don't write for ten people, or two. You write for just one person."