use visible(false) where possible
I grepped the theme for instances of "diffusealpha(0)" and assessed each on a case-by-case basis if I could improve things by adding a call to visible(false).  This typically meant finding actors that I was visibly hiding with diffusealpha(0) that were continuing to draw when they didn't need to be.

Maybe this will be slightly better.  Hopefully it doesn't break anything.

---

One night, while digging through Moonlight's Metrics, I found an inline comment that simply read

# augh, danford

and I lingered on it for a long time, wondering what kind of history went into it.

I realized then that comments were the best part of reading someone's code, because they gave you a sense of what the programmer was *feeling* when they wrote that code.  They were a connection to the human side of the machine.

It was a moment later that I appreciated I would never be a computer scientist.