fix MenuTimer zoom for 4:3 displays
Commit 7439465707c92824961d8ca4482648174fb9f38f fixed the MenuTimer to be more consistently zoomed with the rest of the text in the header, but only checked 16:9.  The sizing of text in the header should be consistent for both 4:3 and 16:9 now.

---

But in their hands I glimpse their souls. Hands are naive, usually: unconsciously curling, or tending to things with a kind of earnest helpfulness it's impossible not to find endearing. I can't even describe it as earnest, because it is devoid of attitude, hands do what hands do, even when they are attached to the cuffs of a wicked man. Watching their movements I know a person had to be there to learn to do those things.