Update ChartParser and MeasureCounter text.
- Countdown the measures left instead of displaying it as a X/Y value.
- Changed some logic to handle breaks better. If there are streams that
are less than measureSequenceThreshold within the break, we separate the
break into chunks.
- Change the criteria for what we consider a measure as "stream".
Before, we would discount if a measure was considered stream if it
contained sequential notes that were too far apart. For example, a
measure containing eight 12th notes and eight 24ths would not be
considered 16th stream. This made We now just look at the number of
notes in a measure.