move StepStats density histogram to ./Scripts
I've factored the Step Statistics density histogram out of ScreenGameplay underlay and into ./Scripts/SL-Histogram.lua where it can be used more easily by different screens for different purposes.  Many of the features that were baked in before are less so, now.

Screens can call NPS_Histogram(), passing in a player object, width, and height, and be returned an ActorMultiVertex histogram representing the NPS density of the current song.  The AMV includes functionality for automatically updating itself in CourseMode when the current song changes.

If a screen wants to scroll the AMV, that is now the screen's responsibility.  (Simply Love's Step Statistics does still handle this for long songs.)

I don't like that NPS_Histogram() makes use of the SL table because it ties the entire file to Simply Love.  I'll think about this some and hopefully revisit it soon.

In general, this would all benefit from better error handling and more comments, but it'll have to do for now.

---

Which would you choose: a world with pyramids or a world without?

"What do you mean?"

Humanity has always dreamt of flying, but the dream is cursed. My aircraft are destined to become tools for slaughter and destruction.

"I know."

But still, I choose a world with pyramids in it. Which world will you choose?

"I just want to create beautiful airplanes."