opt-in nps graph above notefield in gameplay
This adds an option to "Gameplay Extras" to display an NPS density graph above the notefield in gameplay.  A semi-transparent Quad will progressively cover this histogram as gameplay continues to indicate how much of the song has elapsed/is remaining, as made popular by Rhythm Horizon.

The histogram occupies the same space the score normally does, so if this is enabled, the score (usually) won't be shown.

This option is not available in StomperZ because gameplay's header is too small there (and StomperZ should be deprecated/removed).

If this is enabled and Step Statistics is also enabled, the score will be moved over the Step Statistics pane.

If both players have this histogram enabled and they are playing different charts with different peak NPS values, the histogram with the lower peak nps will be vertically scaled so that the two histograms can coexist in the same coordinate space.

---

About 10 years ago or so I locked myself away in a house on the ocean, and I tried to... I said I was trying to write some music, some of which wound up on The Fragile. But what I was really doing was trying to kill myself.

And the whole time I was away by myself, I managed to write one song, which is this song. So when I play it I feel pretty weird about it, because it takes me back to a pretty dark and awful time in my life. It's weird to think how different things are now: I'm still alive, I haven't died yet.