cleanup combo milestones, make hide-able
This cleans up the Lua used in both "Combo milestone" files, making use of visible(false) so that they aren't drawing when they're not needed.

This also adds another toggle to the Hide OptionRow for hiding the combo milestones altogether.  I'm referring to it as "Combo Explosions" because that is what people commonly refer to it as, if they are aware it exists.  This toggle can be saved to player profile.

I added a few strings to en.ini that were accidentally left out of previous commits.

---

In his liner notes and broadcasts, Gould created more than two dozen alter egos for satirical, humorous, and didactic purposes, permitting him to write hostile reviews or incomprehensible commentaries on his own performances. Probably the best-known are the German musicologist Karlheinz Klopweisser, the English conductor "Sir Nigel Twitt-Thornwaite", and the American critic Theodore Slutz. These facets of Gould, whether interpreted as neurosis or "play", have provided ample material for psychobiography.