update fr.ini
Merci Rems.

This also fixes a few issues with text I discovered while testing the new French strings.

---

I clung to nothing, in a way I was calm. But it was a horrible calm–because of my body; my body, I saw with its eyes, I heard with its ears, but it was no longer me; it sweated and trembled by itself and I didn't recognize it any more. I had to touch it and look at it to find out what was happening, as if it were the body of someone else.