improve performance on ScreenSelectProfile
Commit 435e683a29fc761266d554c316ba70d2d7df4069 added player profile data to ScreenSelectProfile.  I didn't appreciate how inefficient my code was until I tested it on a different computer with 20 local profiles – the screen hung for 20+ seconds.  This is obviously unacceptable.

Removing the function that calculated a profile's most frequently earned grade brought performance to "instantaneous" on the computer that had previously taken ~20 seconds.  So, this feature will not be returning until I can improve its runtime by, eh, 5 orders of magnitude or so.

Since both players are sharing the same list of local profiles, I only needed to being retrieving and processing the list once, not once for each player.  That's implemented now, so it should improve the previous runtime by a factor of 2.

The UI for this is not particularly pretty right now, but it functions and I think seeing some data is an improvement overall.  If there is time before the impending release, I'll revisit the layout, otherwise I am okay with it shipping like this in a few days.  I need to move on to the remaining things that need to be fixed and every fix always takes so much longer than I want it to.

---

We are so far, far, very far apart, but it might be that thoughts can overcome time and distance.  Have you ever thought about something like that?