clean up Decents/WayOff code
It was pointed out that "If you're in ECFA mode, 'disable decents and wayoffs' actually disables greats and decents."  This was true; OptionRows were never meant to be flexible with what choices they present.

I've done away with the DecentsWayOffs OptionRow and rebranded the underlying idea as WorstTimingWindow.

It's a little less reliant on clumsy string comparison, a little more flexible with regard to turning off ranges of TimingWindows (can't wait for someone to implement "MGD3" as a Simply Love GameMode), and it adapts the choices it presents based on the current GameMode and language.  Neat.

---

eu escrevi com carinho e atenção. eu to curiosa pra saber oq vai acontecer.