revise SL Acknowledgments
This is the last bit of changes I really strongly wanted to include for this release.  Anything between now and tomorrow will be fluff.

---

"Axl, A thought came to me about it this morning, just as I was waking."

"What thought was that, princess?"

"It was just a thought.  That perhaps God is angry about something we've done.  Or maybe he's not angry, but ashamed."

"A curious thought, princess.  But if it's as you say, why doesn't he punish us?  Why make us forget like fools even things that happened the hour before?"

"Perhaps God's so deeply ashamed of us, of something we did, that he's wishing himself to forget."