Eval Pane4: don't draw offset histogram if no data
Bran has been having an issue with Eval Pane4's offset histogram after upgrading to SL 4.8.3 where the histogram draws as enormously large when no offset data exists (when every judgment was a Miss).

Although I'm completely unable to recreate this, it caused me to reexamine the code involved.  I concluded I was doing a lot of unnecessary work when no offset data was available and I've cleaned up the code accordingly.

Now, if no offset data is available, we don't bother trying to smooth non-existent data, we don't add an empty AMV to the screen, we don't bother calculating mean/median/mode, etc.

This is a good thing to do regardless, and maybe it will resolve whatever is going on with Bran's setup.