check TimingWindowScale before QR code
More work should probably be done to validate TimingWindows before showing the QR code, but this is a start.

Open source StepMania will never be 100% hardened (or even 50%), so it's hardly worth trying, but we should still make an effort to prevent well-intentioned but unknowing users from accidentally submitting invalid scores to GrooveStats.