fix input_redirected bug in SSM's SortWheel
It was previously possible to spam MenuButton input on SelectMusic in just the right way so as to both activate the SortMenu and trigger the engine event that advances the screen to PlayerOptions or Gameplay.

Part of activating the SortMenu is redirecting player input away from the engine into a Lua input callback.  Normally, input handling is returned to the engine when the SortWheel closes.  This would never occur in the scenario I just discovered, causing players to eventually get stuck at ScreenEvaluation, where they would be unable to advance to the next screen without working MenuButtons.

It's really interesting that no one ever reported this bug and I never found it until today.  I'm glad it's fixed now.