make MeasureCounter "rest" counts opt-in
With 4.8.3, some players requested that the "break" or "rest" counter be optional.  Others requested the ability to center the MeasureCounter (which was actually the default behavior, suggesting the UI was unclear).

This commit changes MeasureCounterPosition to be MeasureCounterOptions internally.

For user-facing English strings:

"Measure Counter Position" is now "Measure Counter Options"
"Left" is now "Move Left"
"Up" is now "Move Up"
"Hide Rest Counts" is now a third option in the optionrow

I cleaned up some of the Metrics involved in option explanations for ScreenPlayerOptions.  Text should wrap more naturally, align more cleanly, and be less likely to go out of bounds now.