force more Preferences via SL_Init.lua
Commit dd1157a33b19d5538aa0e2b3e7ee549be66b3f39 added a new screen which prompts users to reset some of the settings in Preferences.ini when they go to switch out of SL to another theme.  You can read more about the rationale in that commit message.

With that in place, I feel more comfortable having SL enforce (silently modify) more global Preferences now.

PercentageScoring should be enabled because ITG scoring depends on it.

AllowW1 should be set to "Everywhere" because that's how ITG did it.  (Even modern DDR does it this way now, too.)

SubSortByNumSteps should be enabled because that's how ITG did it. (There wasn't a Preference for it then; it was hardcoded engine behavior used in sub-sorting the MusicWheel for SortOrders that weren't "Group".)