don't prompt the user if switching to another SL
Commit dd1157a33b19d5538aa0e2b3e7ee549be66b3f39 added an intermediate screen when users would try to switch from SL to any other theme.  The idea was to inform users about global Preferences and offer them the ability to change them to defaults.

I still feel this is a good idea, overall, but some users found it confusing and unnecessary when they were switching from one version of SL to another.  I didn't initially appreciate that people would have multiple different copies of SL installed, but this seems to be quite common in some sub-communities.

This commit adds some naive methods to try to check if the theme being switched to is also SL.  If so, we don't prompt the user about Preferences, we just switch the theme.