deprecate ThemePref for ShowGradesInMusicWheel
People keep asking how to turn on Letter Grades in the MusicWheel suggesting they aren't familiar with the Simply Love Options screen.  I'm opting to deprecate the ThemePref and have the feature always be enabled (when local/USB profiles are being used, of course).

As always, there are too many OptionRows, so any excuse to remove one without eliciting torches and pitchforks is welcome.