slightly better text wrapping algorithm
The previous line break algorithm was very naive and didn't handle combinations of [romance characters, jp characters, emoji] in a single string very well.

This is better now, but there's room for improvement.  For example, in ja.ini, the explanation text for Language contains the substring:

現在、Simply Loveは以下の言語を公式にサポートしています

This has too many characters, so we'll need to insert a line break somewhere.  Ideally, the algorithm should produce something like

現在、Simply Loveは以下の言
語を公式にサポートしています

but it currently splits on the white space between "Simply" and "Love", producing

現在、Simply
Loveは以下の言 語を公式にサポ
ートしています

I'll work on this more soon.