graphic fixes for ScreenTitleMenu
This includes minor fixes for pixel aberrations that occur with on ScreenTitleMenu, for example the "H" in "SIMPLY THONK."

This is an issue with the Wendy font itself, and becomes especially noticeable at larger font sizes. It needs to be fixed manually after rasterizing the font layer. (I discovered this long ago when I printed a custom-made marquee for my machine using Wendy.)