it seems both impossible and necessary
The nightmares are one thing. He can pin you down and suck the skin off your body until your screams wake you, but it ends there. The adrenaline dissipates rapidly and so, too, does the terror.

It is the dreams in which you want him, in which you have him, that leave you in a stupor of sadness that lingers all day.