fix ScreenSelectProfile for USB memorycards
SL 4.8.5 added support for not being forced to use a local profile on ScreenSelectProfile as this feature has been requested since at least v4.8.0.

That (as far I could tell) required me to detangle _fallback's ActorScroller from its input handling from the way it set/unset local profiles and I opted to just swap the entire thing out for a Lua-based sick_wheel scroller.  Three days in, I got a no-profile choice working and failed to remember to check USB profiles in my bleary-eyed joy.

I had, of course, broken support for USB memorycards on ScreenSelectProfile.  This commit fixes it to work as well as it did before.  I could probably try to add more logic to handle (memorycard is present but there was card read error) but this is good enough for now.