Respect BGBrightness when StepStatistics are on
Currently the background is displayed at full brightness when the step
statistics are on, which can be quite distracting.

It took me quite a while to find the right diffuse value (couple of
hours of reading SM source and trying stuff out), but it seems to
visually match with what the StepMania BG implementation does. I
compared some screenshots with step statistics on and off at different
brightness values.

I usually play with HideSongBG, but I noticed that this is wrong, so I
wanted to fix it. It wasn't trivial for me, so I thought it might be
worth sharing it. ;)