fix a few holiday cheer issues
Sometimes the snow wouldn't appear because the table particles references hadn't initialized yet and was empty.  That should be fixed.

I moved the "holiday cheer check" into a global function in ./Script/SL-Helpers.lua rather than recreating the logic everywhere, and then used that function.

I also changed the color of a few BitmapText actors in the event of Holiday Cheer + RainbowMode.