include TimingWindowScale for timing graphs' color
This modifies the DetermineTimingWindow() function in SL-Helpers so that it factors in TimingWindowScale, a global StepMania preference.

This bug was pointed out here: https://twitter.com/HamletEJ/status/1198812583282659328

"Theres a bug where SL’s timing plot/histogram is colored based on the default timing, not the reduced one"

DetermineTimingWindow() is used to to color both the scatter plot and offset histogram on Evaluation.  To be clear, the StepMania engine still assesses judgments during gameplay, and that's working as well (or not) as it always has been.  The themeside change in this commit only impacts coloring the data in SL's Evaluation graphs.

---

The order of operations I used follows along with the relevant src in Player.cpp:

float Player::GetWindowSeconds( TimingWindow tw )
{
	float fSecs = m_fTimingWindowSeconds[tw];
	fSecs *= m_fTimingWindowScale;
	fSecs += m_fTimingWindowAdd;
	return fSecs;
}

---

That is:
original_window * TimingWindowScale + TimingWindowAdd