update GrooveStats URL in QR code to use https
It looks like groovestats.com supports ssl, so I'm opting to update the URL string used in the QR code to be https instead of http.

Using http is working fine for now, but I'm concerned mobile browsers might put http://groovestats.com behind warnings in the future, so this change attempts to preempt that.