clean up SpeedMod display in PlayerOptions
It was demonstrated that rounding errors could be introduced into the SpeedMod helper text that appears at the top of ScreenPlayerOptions, resulting in players with the same x-mod displaying different products.

I went to fix that and cleaned up quite a bit of curious (i.e. bad) code along the way.

1. Wrapping SpeedMods (if the value goes beyond the upper bound or hits 0) now uses modulo operations instead of branching, and the result is then rounded to the nearest appropriate increment.  Explicitly defining a lower bound per-SpeedModType is no longer necessary.

2. I was previously looking up the BPM of the current song every time a player changed their SpeedModType.  It is now looked up once at init.

3. I clarified some unhelpful/vague comments.

Though my intent was to fix a bug, these changes assuredly broke something else I have not yet detected, so I apologize in advance for when it is found and wrecks someone's set.