tidy en.ini
This adds more inline comments to en.ini in the hopes of making things easier for future localization efforts.  I tried to provide context for obscure text elements that might otherwise seem in vacuum where possible.

There's still a lot of font cleanup that needs to be done before the Chinese and Russian localization efforts can be merged in.  I hope to start tackling that soon.