change NPS Histogram colors
These blue/purple values were provided by mute, who got the idea from keize.

---

I'm intrigued by the idea of defining and using a more nuanced, data-meaningful set of colors in the future.  Right now, the bottom of the histogram will always be blue and the top will always be purple, regardless of what the NPS numbers are like.

Would it be helpful to define a standard for the post-ITG community here?  Something like:

 < 6  nps = dark purple
 6-8  nps = purple
 8-10 nps = purple/blue
10-12 nps = blue
12-14 nps = blue/pink
14-16 nps = pink
16-18 nps = pink/orange
18-20 nps = orange
20-22 nps = orange/red
 > 22 nps = red

Obviously I don't know anything about color theory, so those^ values were just made up to illustrate the idea for the sake of this commit message.  I'd want to consult with players who use the graphs (especially colorblind players!) as well as people who understand color theory, but it seems like an interesting possibility.