minor cleanup for MusicWheelItem Grade Lua
I spent a day investigating whether I could build backwards-compatible support for "quint-stars" into FA+ that would "just work" with existing ECFA-Stats.xml files that exist in the wild.

It's possible if we:

1. modify MusicWheelItem.cpp's SetGrade message to include the entire HighScoreList vector as a parameter, and
2. let the theme use that to check for the presence of a "StageAward_FullComboW1" StageAward in the player's HighScore.

but ultimately concluded it wasn't worth it.

Anyway, here's some cleanup that occurred along the way and was worth keeping.