let players know why QR codes aren't appearing
GrooveStats QR code support was added for 4.8.3 as part of a (literally) insane code sprint where my TODO list only grew and grew as the release date remained static for the sake of Stamina RPG.

Part of that rush meant shipping with features that functioned but were not clearly explained.  The QR Code handling included many requirements specified by Archi and if any of them failed, the QR code would not appear.  This understandably lead to many posts such as https://twitter.com/VGR_Okami/status/1190636348249645056

This commit changes the behavior of the QR Code pane so that it
• always appears
• will be covered with a red X if the score is invalid
• will provide a reason (or reasons) why the score was invalid

So far only English is text is provided here.  I'll need to reach out to the translators prior to the next release.