remove support for score vocalization
It's possible that I'll revisit this feature in the future, but I'd need to reach out to each person that contributed content for the original SL3.95 release and see if they are still okay with it eight years later in 2020.

I do not have the time or interest to pursue that right now.