SLAcknowledgments.defrag()
This shuffles the items in SL Acknowledgments around a little in an effort to
keep walls-of-text minimal and overall page count down.

mr.choices.jpg is definitely not legible at 640x480 (or even 800x600) but I'm
including it on the basis that being able to read it is not critical to the
theme's functioning.