more rigorous check if CurrentGame is supported
It was previously possible to:
• use SL's "System Options" UI to switch the game to an unsupported game
  (lights, beat, kickbox, etc.)
• receive a warning that the game you just switched to is unsupported
  and that you need to switch to a different game
• ignore the warning and proceed to play the unsupported game, probably getting
  yourself quickly stuck in a broken screen

This makes better use of SL's global CurrentGameIsSupported() function by not
allowing players to leave the operator menu in an unsupported game.

... while testing game switching, I found that SL's global BrighterOptionRows()
function was throwing Lua errors stemming from the _fallback theme's ThemePrefs
system.  I rebranded "BrighterOptionRows()" to "DarkUI()", used it in a few
more places throughout SL, added inline comments explaining the issue, and
effectively worked my way around it.  The issue should really be patched in
_fallback.