right-align MusicWheelItems (...kind of)
At certain resolutions and aspect ratios, a single pixel gap would appear between the right side of the MusicWheel and the right edge of the screen.  I've been aware of this for a while and assumed it was due to rounding errors because of how badly I'd originally written all the code for ScreenSelectMusic back in 2013.

This only happened at certain resolutions and, to be honest, was only really apparent when RainbowMode was enabled, so I gave it low priority on my list of things to fix.

I set out to fix it today, thinking "easy, I'll just right-align the MusicWheelItems and set their x value to SCREEN_WIDTH."

I'm left with messy, ad-hoc code I am no less embarrassed by that (sigh) does accomplish the goal of eliminating that 1 pixel gap that sometimes appeared.  Spoilers: it accomplishes this by extending the overall width of the MusicWheelItems giving each a bit of overflow that draws beyond the right edge of the screen.

Eyeballing the before and after in 4:3, 16:9, and 16:10, I can say with some assurance that, after several hours of fiddling with numbers and alignment and zoom values, users should not see any difference in the way ScreenSelectMusic looks beyond the 1px gap no longer happening.

It still passes the "How To Cook Delicious Rice and the Effects of Eating Rice" test, so that's something.