handle split BPMs in PlayerOptions's MusicRate row
Previously, if the current song had divergent timing data and P1 and P2 were trying to choose stepcharts with different steps timing, the MusicRate OptionRow would just use the steps timing from the MasterPlayerNumber (typically PLAYER_1).

With this commit, the bpm range in the MusicRate OptionRow gutter will now read "(split)" in such a scenario.  This is not ideal, and hopefully just a placeholder for a little while.