update Bebas Neue ComboFont
The % glyph is much wider than the other glyphs here.  I'd previously tried to be
lazy and just squish it to fit in 54e68fc1ab8246999c72d2f86fe493470ae8d03d; but that looked bad.

It was necessary to re-layout the overall sprite sheet.  My Affinity/Photoshop
strategy for now is to specify:

• 800x200 pixel canvas
• 100pt characters that are each their own layer/object
• characters are vertically aligned to middle and horizontally aligned to center
• characters will be nudged on a per-character basis from there if needed
  (glyphs in the Work font are not consistently vertically aligned?)