use steps:GetFilename() in stepchart parsing
I found that steps objects have a GetFilename() method that "Returns the
Steps filename from the Cache."

https://quietly-turning.github.io/Lua-For-SM5/LuaAPI#Actors-Steps-GetFilename

I'd previously been recreating this in-theme with a combination of the
filepath to the current song and some assumptions that ssc files should
be preferred over sm files for parsing if both are found.  It makes more
sense to use whatever file (ssc or sm) the cache file specifies as that
is (hopefully!) where gameplay will get its steps from.

This changes GetSimfileString() in SL-ChartParser.lua to use
steps:GetFilename() and updates other functions that interact with it
accordingly.