bug with autostyle when starting as p2
Currently if you start the game as P2 (and P1 not joined) with autostyle set to single, the branch will unjoin you and then proceed to ScreenSelectPlayMode with no players so some errors are thrown. Looking at the comments, it looks like it's only supposed to unjoin if two people are playing so i just added that in to the if statement.