add support for ddr difficulties + ssc hashes
currently when creating a hash in SL-HashGenerator or a simfilestring in SL-ChartParser it uses steps:GetDifficulty() to check for the correct difficulty in the actual file. some files (maybe older ones especially?) use ddr difficulties (light/standard/heavy/expert) which will not match what GetDifficulty returns so hashes/streams are never done.

this creates a table so that, for example, expert or challenge will be equivalent.

additionally, for hashes, the current hashgenerator looks for a notes table with 7 or more entries and specific things on certain entries. SSC files are arranged a little differently. this adds support for them.

caveat: SL-ChartParser still won't work for SSC files with ddr difficulties