support for old difficulties + ssc hashes
currently when creating a hash in SL-HashGenerator or a simfilestring in SL-ChartParser stepmania difficulties are assumed (challenge,hard,medium,easy) to check for the correct difficulty in the actual file. some files (maybe older ones especially?) use ddr difficulties (light/standard/heavy/expert) which will not match so hashes/streams are never done.

additionally, for hashes, the current hashgenerator looks for a notes table with 7 or more entries and specific things on certain entries. SSC files are arranged a little differently. this adds support for them.

caveat: SL-ChartParser still won't work for SSC files with ddr difficulties because that uses match to look for the difficulty in the files so a simple conversion won't fix it as we're going the other way (standard difficulties -> old ones)